pyproject.toml
[tool.poetry]
name = "klassbot"
version = "0.2.1"
description = ""
authors = ["hexatester <revolusi147id@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^12.7"
toml = "^0.10.1"
SQLAlchemy = "^1.3.17"
python-i18n = {extras = ["yaml"], version = "^0.3.7"}
python-dateutil = "^2.8.1"
psycopg2-binary = "^2.8.5"
sqlalchemy-utils = "^0.36.6"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^19.10b0"
flake8 = "^3.8.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
